+++ dlink-web/src/components/Studio/StudioEvent/DDL.ts
@@ -3,6 +3,22 @@ import FlinkSQL from "./FlinkSQL";
 import {SessionType} from "@/pages/FlinkSqlStudio/model";
 import {Modal,message} from "antd";
 import {addOrUpdateData, getData, handleRemove} from "@/components/Common/crud";
+
+/*--- 保存sql ---*/
+
+export function saveTask(current:any,dispatch: any){
+  if (current.task) {
+    let task = {
+      ...current.task,
+      statement: current.value,
+    };
+    dispatch && dispatch({
+      type: "Studio/saveTask",
+      payload: task,
+    });
+  }
+}
+
 /*--- 创建会话 ---*/
 export function createSession(session: SessionType,dispatch: any) {
   const res = addOrUpdateData("api/studio/createSession",session)
@@ -32,6 +48,7 @@ export function changeSession(session: SessionType, dispatch: any) {
     showTables(session.session,dispatch);
   },200);
 }
+
 /*--- 退出会话 ---*/
 export function quitSession( dispatch: any) {
   dispatch && dispatch({
