+++ dlink-web/src/pages/FlinkSqlStudio/model.ts
@@ -92,6 +92,7 @@ export type TabsItemType = {
   task?: TaskType;
   console: ConsoleType;
   monaco?: any;
+  isModified:boolean;
   sqlMetaData?:SqlMetaData;
 }
 
@@ -179,6 +180,7 @@ const Model: ModelType = {
       value: '',
       closable: false,
       path: ['草稿'],
+      isModified: false,
       task: {
         jobName: '草稿',
         // type: 'standalone',
@@ -215,6 +217,7 @@ const Model: ModelType = {
         key: 0,
         value: '',
         closable: false,
+        isModified: false,
         path: ['草稿'],
         task: {
           jobName: '草稿',
@@ -301,10 +304,12 @@ const Model: ModelType = {
       let newTabs = state.tabs;
       if(newCurrent.key == payload.activeKey){
         newCurrent.sqlMetaData = payload.sqlMetaData;
+        newCurrent.isModified = payload.isModified;
       }
       for (let i = 0; i < newTabs.panes.length; i++) {
         if (newTabs.panes[i].key == payload.activeKey) {
           newTabs.panes[i].sqlMetaData = payload.sqlMetaData;
+          newTabs.panes[i].isModified = payload.isModified;
           break;
         }
       }
@@ -325,6 +330,7 @@ const Model: ModelType = {
         ...state,
         current: {
           ...newCurrent,
+          isModified:false,
         },
         tabs: {
           ...payload,
